begin ts-esbuild-react as a CLI module
