+++ src/index.ts
@@ -3,11 +3,11 @@
 
 import CLI from "./CLI";
 
-const main = async (argv: string[]) => {
+export const main = async (argv: string[]) => {
   const cli = new CLI();
   const args = await cli.process(argv);
   console.info(args);
   return;
 };
 
-export = (async () => await main(process.argv))();
+export default main;
